test/test_app.py
# Create new clean test file
import pytest
import sys
import os

# Add the parent directory to Python path
sys.path.insert(0, os.path.abspath('.'))

# Import from the root directory
import __init__
from __init__ import create_app, db


@pytest.fixture
def app():
    app = create_app()
    app.config['TESTING'] = True
    app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///:memory:'
    app.config['WTF_CSRF_ENABLED'] = False
    
    with app.app_context():
        db.create_all()
        yield app
        db.drop_all()


@pytest.fixture
def client(app):
    return app.test_client()


def test_health_check(client):
    response = client.get('/health')
    assert response.status_code == 200
    assert response.json['status'] == 'healthy'


def test_home_page(client):
    response = client.get('/')
    assert response.status_code in [200, 302]  # 200 or redirect


def test_register_page(client):
    response = client.get('/auth/register')
    assert response.status_code == 200


def test_login_page(client):
    response = client.get('/auth/login')
    assert response.status_code == 200


def test_basic_coverage():
    """Basic test for coverage"""
    assert 1 + 1 == 2
