requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-JWT-Extended==4.5.3
Werkzeug==3.0.1  # UPDATED - fixes security issues
WTForms==3.1.2
python-dotenv==1.0.0
pytest-cov==4.1.0